Add crackable module. All tests still passing
